+++ stm32/jt/TouchGFX/gui/src/setamperescreen_screen/setAmpereScreenPresenter.cpp
@@ -22,7 +22,7 @@ void setAmpereScreenPresenter::storeWeldingAmps(float ampsV)
 	model->storeWeldingAmps(ampsV);
 }
 
-uint32_t setAmpereScreenPresenter::getWeldingAmps()
+float setAmpereScreenPresenter::getWeldingAmps()
 {
 	uint32_t res = 0;
 	res = model->getWeldingAmps();
